lambda-client/requirements.txt
werkzeug
json-rpc
jsonrpc-requests
cachetools
schedule
requests
python-dotenv
flask
github-webhook
